redirects, some error enforcement on uploads... some error with big files

+++ src/config/debug.cpp
@@ -74,7 +74,7 @@ void ConfigHandler::printRegisteredConfs(std::string filename, std::string pwd)
 		}
 	};
 
-	auto printSize = [](long& value, const std::string& label,
+	auto printSize = [](unsigned long long& value, const std::string& label,
 					const std::string& default_value = "1m", int padding = 30) {
 		Logger::white(label, false, padding);
